requirements.txt
requests
python-dotenv
azure-ai-inference
azure-core
bs4